chore: Update install, uninstall, and update scripts
Updated `install_pieye.sh` to include the option of passing a custom repository URL as the first argument. Additionally, added error checks to ensure that Git and Pip are installed on the system. The script now checks for the existence of the repository directory before cloning and installing the Pieye camera software. It enables the watchdog and sets up auto-login on the command line. Moreover, it copies `display_info.sh` to the `/home/pi` folder and adds an entry to execute `display_info.sh` on login.

In the `uninstall_pieye.sh` script, removed `display_info.sh` from the `/home/pi` folder and eliminated the corresponding entry from the `.profile` file.

In the `update_pieye.sh` script, updated the logic to check for local changes before updating the Pieye camera software, preventing updates when uncommitted changes exist. Error handling ensures that the Pieye service starts correctly after the update.