feat: Add setup_pi_preboot.sh
This commit introduces the setup_pi_preboot.sh script, which automates the setup process for Raspberry Pi devices before booting. The script modifies the config.txt and cmdline.txt files, enables SSH, and sets the MAC address for USB-Ethernet functionality. This automation simplifies the initial setup of Raspberry Pi devices. This makes it quicker to setup the devices.